% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{generateDistn}
\alias{generateDistn}
\title{Generate theoretical GC content distributions}
\usage{
generateDistn(seqs, genome, nchrom, file = "fastqc_theoretical_gc.txt",
  n = 1e+06, bp = 100, wts = 1, name = "")
}
\arguments{
\item{seqs}{a DNAStringSet of the sequences to simulate
read from. E.g. for RNA-seq, the transcripts, which
can be generated with \code{extractTranscriptSeqs}
from the GenomicFeatures package.
See the example script located in \code{inst/script/human_mouse.R}}

\item{genome}{a BSgenome object.
See the example script located in \code{inst/script/human_mouse.R}}

\item{nchrom}{the number of chromosomes from the genome to simulate
reads from}

\item{file}{the path of the file to write out}

\item{n}{the number of reads to simulate}

\item{bp}{the basepair of the reads}

\item{wts}{optional weights to go along with the \code{seqs} or
the chromosomes in \code{genome}, e.g. to represent
more realistic expression of transcripts}

\item{name}{the name to be printed at the top of the file}
}
\value{
the name of the file which was written
}
\description{
This function generates random simulated reads from
either provided \code{seqs} (best for RNA-seq)
or from a genome (best for DNA-seq). The GC content
of these reads is then tabulated to produce a distribution
file which can be read by MultiQC to be displayed
on top of the FASTQC GC content module. Either \code{seqs}
or \code{genome} is required, and only one can be specified.
Specifying \code{genome} requires also specifying \code{nchrom}.
}
\references{
MultiQC: 
http://multiqc.info/

FASTQC:
http://www.bioinformatics.babraham.ac.uk/projects/fastqc/
}

